+++ ARCL/Source/Nodes/PolylineNode.swift
@@ -10,7 +10,6 @@ import Foundation
 import SceneKit
 import MapKit
 
-
 /// A block that will build an SCNBox with the provided distance.
 /// Note: the distance should be aassigned to the length
 public typealias BoxBuilder = (_ distance: CGFloat) -> SCNBox